+++ R/gt_img_circle.R
@@ -1,4 +1,4 @@
-#' Create a circular border around a
+#' Create a circular border around a image
 #'
 #' @param value The source image
 #' @param height The height in pixels of the circle
@@ -20,17 +20,89 @@ img_circle <- function(value, height, border_color, border_weight) {
   image
 }
 
-#' Create circular border around an image
+#' Create a square colored border around an image
+#'
+#' @param value The source image
+#' @param height The height in pixels of the circle
+#' @param border_color A string indicating the color of the border
+#' @param border_weight The weight of the border in pixels
+#'
+#' @return HTML
+
+img_square <- function(value, height, width, border_color, border_weight) {
+  image <- htmltools::div(
+    style = glue::glue(
+      "background-image: url({value});background-size:cover;",
+      "background-position:center;",
+      "border-bottom: {border_weight}px solid {border_color};",
+      "border-radius: 0%;height:{height}px;width:{width}px;",
+      "object-fit: contain;"
+    )
+  )
+
+  image
+}
+
+#' Create an identifier line border at the bottom of an image
 #'
 #' @param gt_object An existing gt object
 #' @param column The column to apply the transformation to
 #' @param height A number indicating the height of the image in pixels.
-#' @param border_color The color of the circular border, can either be a single value ie (`white` or `#FF0000`) or a vector where the lenght of the vector is equal to the number of rows.
+#' @param border_color The color of the circular border, can either be a single value ie (`white` or `#FF0000`) or a vector where the length of the vector is equal to the number of rows.
 #' @param border_weight A number indicating the weight of the border in pixels.
 #' @return a gt object
 #' @export
 #'
 #' @section Examples:
+#'
+#' ```r
+#' library(gt)
+#' gt_img_tab <- dplyr::tibble(
+#'   x = 1:4,
+#'   names = c("Waking Up",  "Wiggling", "Sleep"," Glamour"),
+#'   img = c(
+#'      "https://pbs.twimg.com/media/EiIY-1fXgAEV6CJ?format=jpg&name=360x360",
+#'      "https://pbs.twimg.com/media/EiIY-1fXcAIPdTS?format=jpg&name=360x360",
+#'      "https://pbs.twimg.com/media/EiIY-1mX0AE-YkC?format=jpg&name=360x360",
+#'      "https://pbs.twimg.com/media/EiIY-2cXYAA1VaO?format=jpg&name=360x360"
+#'   )
+#' ) %>%
+#'   gt() %>%
+#'   gt_img_border(img)
+#' ```
+#' @section Figures:
+#' \if{html}{\figure{gt_img_circle.png}{options: width=80\%}}
+#'
+#' @family Utilities
+gt_img_border <- function(gt_object, column, height = 25, width = 25,
+                          border_color = "black", border_weight = 2.5){
+
+  stopifnot("'gt_object' must be a 'gt_tbl', have you accidentally passed raw data?" = "gt_tbl" %in% class(gt_object))
+
+  gt_object %>%
+    text_transform(
+      locations = cells_body({{column}}),
+      fn = function(value){
+        mapply(img_square, value, height, width, border_color,
+               border_weight, SIMPLIFY = FALSE)
+      }
+    )
+}
+
+
+#' Create circular border around an image
+#'
+#' @param gt_object An existing gt object
+#' @param column The column to apply the transformation to
+#' @param height A number indicating the height of the image in pixels.
+#' @param border_color The color of the circular border, can either be a single value ie (`white` or `#FF0000`) or a vector where the length of the vector is equal to the number of rows.
+#' @param border_weight A number indicating the weight of the border in pixels.
+#' @return a gt object
+#' @export
+#'
+#' @section Examples:
+#'
+#' ```r
 #' library(gt)
 #' gt_img_tab <- dplyr::tibble(
 #'   x = 1:4,
@@ -44,6 +116,7 @@ img_circle <- function(value, height, border_color, border_weight) {
 #' ) %>%
 #'   gt() %>%
 #'   gt_img_circle(img)
+#' ```
 #' @section Figures:
 #' \if{html}{\figure{gt_img_circle.png}{options: width=80\%}}
 #'